+++ Cindy3D/src/de/tum/in/cindy3dplugin/jogl/primitives/Line.java
@@ -18,7 +18,7 @@ public class Line extends Primitive {
 
 	public Line(double x1, double y1, double z1,
 				double x2, double y2, double z2,
-				double radius, Color color, double shininess,
+				double radius, Color color, int shininess,
 				LineType lineType) {
 		
 		super(color, shininess, 1);
