+++ Cindy3D/src/de/tum/in/cindy3dplugin/jogl/primitives/Mesh.java
@@ -24,7 +24,7 @@ public class Mesh extends Primitive {
 	public int faceCount;
 
 	public Mesh(int height, int width, double[][] positions,
-			double[][] normals, Color color, double shininess, double alpha,
+			double[][] normals, Color color, int shininess, double alpha,
 			MeshTopology topology) {
 		super(color, shininess, alpha);
 		init(width, height, positions, normals, normals != null, topology);
@@ -33,7 +33,7 @@ public class Mesh extends Primitive {
 	}
 
 	public Mesh(int height, int width, double[][] positions,
-			boolean perVertexNormals, Color color, double shininess,
+			boolean perVertexNormals, Color color, int shininess,
 			double alpha, MeshTopology topology) {
 		super(color, shininess, alpha);
 		init(width, height, positions, null, perVertexNormals, topology);
